fix: improve `SRI_REGEX` (#151)
Update `SRI_REGEX` to avoid potential polynomial evaluation time due to
possible exchange of characters between `[^?]+` and `[?\S*]*`. Requiring
a `?` explicitly avoids this problem and making the last group optional
is required to match the previous zero-or-more matching behavior.

This now more closely matches the regex at the end of STRICT_SRI_REGEX.